auction ui directory reorg
