+++ @dao-auction/data/daoAuctionQuery.ts
@@ -1,22 +1,6 @@
 import gql from 'graphql-tag'
 
-/**
- * https://playground.api.zora.co/
- */
-export const TOKEN_METADATA_QUERY = gql`
-  query NounishAuctions($collectionAddress: String!, $tokenId: String!) {
-    token(
-      token: {address: $collectionAddress, tokenId: $tokenId}
-    ) {
-      token {
-        description
-        metadata
-      }
-    }
-  }
-`
-
-export const NOUNISH_AUCTIONS_QUERY = gql`
+export const DAO_AUCTION_QUERY = gql`
   query NounishAuctions($collectionAddress: String!, $network: NetworkInput) {
     nouns {
       nounsActiveMarket(
