+++ @dao-auction/data/daoTokenQuery.ts
@@ -1,0 +1,14 @@
+import gql from 'graphql-tag'
+
+export const DAO_TOKEN_QUERY = gql`
+  query NounishAuctions($collectionAddress: String!, $tokenId: String!) {
+    token(
+      token: {address: $collectionAddress, tokenId: $tokenId}
+    ) {
+      token {
+        description
+        metadata
+      }
+    }
+  }
+`
