+++ @dao-auction/types/zora.api.generated.ts
